tools/rccl-sweep/requirements.txt
# RCCL Test Sweep Tool Dependencies
# Prefer using conda: conda env create -f environment.yml
# This file is for pip fallback if needed

pyyaml>=6.0
pandas>=1.3.0
tabulate>=0.9.0
colorama>=0.4.6
matplotlib>=3.5.0

